feat: 新增注意力权重可视化工具和确定性问答功能
refactor(kb_retriever): 重构知识库检索器支持批量处理
feat(eval_utils): 添加确定性问答函数answer_question_deterministic
feat(metrics): 实现多维度评估指标计算
fix(config): 将kb_scale_factor类型改为float
docs: 更新.gitignore忽略实验输出文件

+++ src/kblam/models/kblam_config.py
@@ -6,7 +6,7 @@ class KBLaMConfig(PretrainedConfig):
         self,
         base_model_name_or_path: str = "",
         kb_layer_frequency: int = 3,
-        kb_scale_factor: int | None = None,
+        kb_scale_factor: float | None = None,
         top_k_kb: int = 100,
         dynamic_sparsify: bool = False,
         sep_query_head: bool = False,
